Delete: 필요없는 파일 삭제 (이전 온보딩 뷰 5개, LocalDataSource)
