--- data/src/main/kotlin/com/acon/data/datasource/local/OnboardingLocalDataSource.kt
@@ -1,11 +1,0 @@
-package com.acon.data.datasource.local
-
-interface OnboardingLocalDataSource {
-    var dislikeFoodList: Set<String>
-    var favoriteCuisineRank: List<String>
-    var favoriteSpotType: String
-    var favoriteSpotStyle: String
-    var favoriteSpotRank: List<String>
-
-    fun clear()
-}
