+++ data/src/main/kotlin/com/acon/data/repository/OnboardingRepositoryImpl.kt
@@ -1,12 +1,10 @@
 package com.acon.data.repository
 
 import android.util.Log
-import com.acon.data.datasource.local.OnboardingLocalDataSource
 import com.acon.data.datasource.remote.OnboardingRemoteDataSource
 import com.acon.data.dto.request.PostOnboardingResultRequest
 import com.acon.data.error.runCatchingWith
 import com.acon.domain.error.onboarding.PostOnboardingResultError
-import com.acon.domain.model.onboarding.OnboardingPreferences
 import com.acon.domain.repository.OnboardingRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -17,7 +15,6 @@ import javax.inject.Inject
 
 class OnboardingRepositoryImpl @Inject constructor(
     private val onboardingRemoteDataSource: OnboardingRemoteDataSource,
-    private val onboardingLocalDataSource: OnboardingLocalDataSource
 ) : OnboardingRepository {
 
     private val _onboardingResultStateFlow = MutableStateFlow<Result<Unit>?>(null)
@@ -57,35 +54,5 @@ class OnboardingRepositoryImpl @Inject constructor(
                 Result.failure(exception)
             }
         }
-    }
-
-    override fun postDislikeFood(choice: Set<String>){
-        onboardingLocalDataSource.dislikeFoodList = choice
-    }
-
-    override fun postFavoriteCuisineRank(choice: List<String>) {
-        onboardingLocalDataSource.favoriteCuisineRank = choice
-    }
-
-    override fun postFavoriteSpotType(choice: String) {
-        onboardingLocalDataSource.favoriteSpotType = choice
-    }
-
-    override fun postFavoriteSpotStyle(choice: String) {
-        onboardingLocalDataSource.favoriteSpotStyle = choice
-    }
-
-    override fun postFavoriteSpotRank(choice: List<String>) {
-        onboardingLocalDataSource.favoriteSpotRank = choice
-    }
-
-    override fun getOnboardingResults(): OnboardingPreferences {
-        return OnboardingPreferences(
-            dislikeFoodList = onboardingLocalDataSource.dislikeFoodList,
-            favoriteCuisineRank = onboardingLocalDataSource.favoriteCuisineRank,
-            favoriteSpotType = onboardingLocalDataSource.favoriteSpotType,
-            favoriteSpotStyle = onboardingLocalDataSource.favoriteSpotStyle,
-            favoriteSpotRank = onboardingLocalDataSource.favoriteSpotRank
-        )
     }
 }
