+++ domain/src/main/java/com/acon/domain/repository/OnboardingRepository.kt
@@ -1,6 +1,5 @@
 package com.acon.domain.repository
 
-import com.acon.domain.model.onboarding.OnboardingPreferences
 import kotlinx.coroutines.flow.StateFlow
 
 interface OnboardingRepository {
@@ -12,14 +11,6 @@ interface OnboardingRepository {
         favoriteSpotStyle: String,
         favoriteSpotRank: List<String>
     ): Result<Unit>
-
-    fun postDislikeFood(choice: Set<String>)
-    fun postFavoriteCuisineRank(choice: List<String>)
-    fun postFavoriteSpotType(choice: String)
-    fun postFavoriteSpotStyle(choice: String)
-    fun postFavoriteSpotRank(choice: List<String>)
-
-    fun getOnboardingResults(): OnboardingPreferences
 
     val onboardingResultStateFlow: StateFlow<Result<Unit>?>
 }
