--- feature/onboarding/src/main/java/com/acon/feature/onboarding/screen/delete/FrequentPlaceSelectScreen/composable/FrequentPlaceSelectScreenContainer.kt
@@ -1,2 +1,0 @@
-package com.acon.feature.onboarding.screen.delete.FrequentPlaceSelectScreen.composable
-
