--- feature/onboarding/src/main/java/com/acon/feature/onboarding/screen/delete/PerferredPlaceRateScreen/PerferredPlaceRateScreenViewModel.kt
@@ -1,78 +1,0 @@
-package com.acon.feature.onboarding.screen.delete.PerferredPlaceRateScreen
-
-import com.acon.core.utils.feature.base.BaseContainerHost
-import com.acon.domain.repository.OnboardingRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import org.orbitmvi.orbit.Container
-import org.orbitmvi.orbit.viewmodel.container
-import javax.inject.Inject
-
-private const val ONBOARDING_TOTAL_PAGES = 5;
-
-@HiltViewModel
-class PreferredPlaceRateScreenViewModel @Inject constructor(
-    private val onboardingRepository: OnboardingRepository
-) : BaseContainerHost<PreferredPlaceRateScreenState, PreferredPlaceRateScreenSideEffect>() {
-
-    override val container: Container<PreferredPlaceRateScreenState, PreferredPlaceRateScreenSideEffect> =
-        container(
-            initialState = PreferredPlaceRateScreenState(  )
-        )
-
-    fun onCardClicked(id: String) = intent {
-
-        val updatedSelectedCard = state.selectedCard.toMutableList()
-
-        if (updatedSelectedCard.contains(id)) {
-            updatedSelectedCard.remove(id)
-        } else if (updatedSelectedCard.size < 4) {
-            updatedSelectedCard.add(id)
-        }
-        reduce {
-            state.copy(selectedCard = updatedSelectedCard.toSet()) // Set으로 변환
-        }
-    }
-
-    fun showDialog() = intent {
-        reduce {
-            state.copy(openCloseDialog = true)
-        }
-    }
-
-    fun hideDialog() = intent {
-        reduce {
-            state.copy(openCloseDialog = false)
-        }
-    }
-
-    fun skipConfirmed() = intent {
-        reduce {
-            state.copy(openCloseDialog = false)
-        }
-        postSideEffect(PreferredPlaceRateScreenSideEffect.NavigateToLastPage)
-    }
-
-    fun navigateToPreviousPage() = intent {
-        postSideEffect(PreferredPlaceRateScreenSideEffect.NavigateToPreviousPage)
-    }
-
-    fun navigateToNextPage() = intent {
-        onboardingRepository.postFavoriteSpotRank(state.selectedCard.toList())
-
-        //만약 local DB에 정보가 다 들어가 있으면.
-        postSideEffect(PreferredPlaceRateScreenSideEffect.NavigateToNextPage)
-    }
-}
-
-data class PreferredPlaceRateScreenState(
-    val selectedCard: Set<String> = emptySet(),
-    val totalPages: Int = ONBOARDING_TOTAL_PAGES,
-    val currentPage: Int = 4,
-    val openCloseDialog: Boolean = false,
-)
-
-sealed interface PreferredPlaceRateScreenSideEffect {
-    data object NavigateToPreviousPage: PreferredPlaceRateScreenSideEffect
-    data object NavigateToNextPage: PreferredPlaceRateScreenSideEffect
-    data object NavigateToLastPage: PreferredPlaceRateScreenSideEffect
-}
